# ReklamacjeAPI/openapi.yaml
openapi: 3.0.3
info:
  title: Reklamacje API - Returns/Warehouse
  version: 1.0.0
servers:
  - url: https://localhost:5001
paths:
  /api/returns:
    get:
      summary: List returns
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: statusWewnetrzny
          schema: { type: string }
        - in: query
          name: statusAllegro
          schema: { type: string }
        - in: query
          name: handlowiecId
          schema: { type: integer }
        - in: query
          name: search
          schema: { type: string }
      responses:
        '200':
          description: Returns list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnListResponse'
  /api/returns/{id}:
    get:
      summary: Get return details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Return details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnDetailsResponse'
  /api/returns/{id}/warehouse:
    patch:
      summary: Update warehouse data for return
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnWarehouseUpdateRequest'
      responses:
        '200':
          description: Updated
  /api/returns/{id}/forward-to-sales:
    post:
      summary: Forward return to sales
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnForwardToSalesRequest'
      responses:
        '200':
          description: Forwarded
  /api/returns/{id}/decision:
    patch:
      summary: Save sales decision
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnDecisionRequest'
      responses:
        '200':
          description: Decision saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnDecisionResponse'
  /api/returns/manual:
    post:
      summary: Create manual return
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnManualCreateRequest'
      responses:
        '201':
          description: Manual return created
  /api/returns/{id}/archive:
    patch:
      summary: Archive return
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Archived
  /api/returns/{id}/actions:
    get:
      summary: List return actions
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Action list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnActionsResponse'
    post:
      summary: Add action to return
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnActionCreateRequest'
      responses:
        '201':
          description: Action created
  /api/returns/summary:
    get:
      summary: Summary of returns
      parameters:
        - in: query
          name: handlowiecId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: dateFrom
          schema: { type: string, format: date }
        - in: query
          name: dateTo
          schema: { type: string, format: date }
      responses:
        '200':
          description: Summary response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReturnSummaryResponse'
  /api/returns/summary/export:
    get:
      summary: Export returns summary
      parameters:
        - in: query
          name: handlowiecId
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: dateFrom
          schema: { type: string, format: date }
        - in: query
          name: dateTo
          schema: { type: string, format: date }
      responses:
        '200':
          description: CSV export
  /api/returns/{id}/forward-to-complaints:
    post:
      summary: Forward return to complaints
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForwardToComplaintRequest'
      responses:
        '200':
          description: Forwarded
  /api/messages:
    get:
      summary: List messages
      parameters:
        - in: query
          name: returnId
          schema: { type: integer }
      responses:
        '200':
          description: Messages list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagesResponse'
    post:
      summary: Create message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageCreateRequest'
      responses:
        '201':
          description: Message created
  /api/messages/{id}/read:
    patch:
      summary: Mark message as read
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Marked read
  /api/warehouse/search:
    get:
      summary: Warehouse search
      parameters:
        - in: query
          name: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Warehouse search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseSearchResponse'
  /api/warehouse/intake:
    post:
      summary: Intake equipment to warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WarehouseIntakeRequest'
      responses:
        '201':
          description: Intake created
components:
  schemas:
    ReturnListItemDto:
      type: object
      properties:
        id: { type: integer }
        referenceNumber: { type: string }
        waybill: { type: string, nullable: true }
        buyerName: { type: string }
        productName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        statusAllegro: { type: string, nullable: true }
        statusWewnetrzny: { type: string, nullable: true }
        stanProduktu: { type: string, nullable: true }
        decyzjaHandlowca: { type: string, nullable: true }
        handlowiecId: { type: integer, nullable: true }
        isManual: { type: boolean }
    ReturnDetailsDto:
      type: object
      properties:
        id: { type: integer }
        referenceNumber: { type: string }
        createdAt: { type: string, format: date-time }
        statusWewnetrzny: { type: string, nullable: true }
        statusAllegro: { type: string, nullable: true }
        buyerLogin: { type: string, nullable: true }
        buyerName: { type: string }
        buyerPhone: { type: string, nullable: true }
        buyerAddress: { type: string, nullable: true }
        waybill: { type: string, nullable: true }
        carrierName: { type: string, nullable: true }
        productName: { type: string, nullable: true }
        offerId: { type: string, nullable: true }
        quantity: { type: integer, nullable: true }
        reason: { type: string, nullable: true }
        uwagiMagazynu: { type: string, nullable: true }
        stanProduktuId: { type: integer, nullable: true }
        stanProduktuName: { type: string, nullable: true }
        przyjetyPrzezId: { type: integer, nullable: true }
        przyjetyPrzezName: { type: string, nullable: true }
        dataPrzyjecia: { type: string, format: date-time, nullable: true }
        decyzjaHandlowcaId: { type: integer, nullable: true }
        decyzjaHandlowcaName: { type: string, nullable: true }
        komentarzHandlowca: { type: string, nullable: true }
        dataDecyzji: { type: string, format: date-time, nullable: true }
        isManual: { type: boolean }
    ReturnWarehouseUpdateRequest:
      type: object
      required: [stanProduktuId, dataPrzyjecia, przyjetyPrzezId]
      properties:
        stanProduktuId: { type: integer }
        uwagiMagazynu: { type: string, nullable: true }
        dataPrzyjecia: { type: string, format: date-time }
        przyjetyPrzezId: { type: integer }
    ReturnForwardToSalesRequest:
      type: object
      required: [stanProduktuId]
      properties:
        stanProduktuId: { type: integer }
        uwagiMagazynu: { type: string, nullable: true }
    ReturnDecisionRequest:
      type: object
      required: [decyzjaId]
      properties:
        decyzjaId: { type: integer }
        komentarz: { type: string, nullable: true }
    ReturnDecisionResponse:
      type: object
      properties:
        returnId: { type: integer }
        statusWewnetrzny: { type: string }
        decyzjaHandlowca: { type: string }
        dataDecyzji: { type: string, format: date-time }
    ReturnManualCreateRequest:
      type: object
      required: [numerListu, stanProduktuId, buyerFullName, wybraniHandlowcy]
      properties:
        numerListu: { type: string }
        produkt: { type: string, nullable: true }
        przewoznik: { type: string, nullable: true }
        stanProduktuId: { type: integer }
        uwagiMagazynu: { type: string, nullable: true }
        buyerFullName: { type: string }
        buyerStreet: { type: string, nullable: true }
        buyerZipCode: { type: string, nullable: true }
        buyerCity: { type: string, nullable: true }
        buyerPhone: { type: string, nullable: true }
        wybraniHandlowcy:
          type: array
          items: { type: integer }
    ReturnActionDto:
      type: object
      properties:
        id: { type: integer }
        returnId: { type: integer }
        data: { type: string, format: date-time }
        uzytkownik: { type: string }
        tresc: { type: string }
    ReturnActionCreateRequest:
      type: object
      required: [tresc]
      properties:
        tresc: { type: string }
    MessageDto:
      type: object
      properties:
        id: { type: integer }
        nadawcaId: { type: integer }
        odbiorcaId: { type: integer }
        tytul: { type: string, nullable: true }
        tresc: { type: string }
        dataWyslania: { type: string, format: date-time }
        dotyczyZwrotuId: { type: integer, nullable: true }
        czyPrzeczytana: { type: boolean }
    MessageCreateRequest:
      type: object
      required: [nadawcaId, odbiorcaId, tresc]
      properties:
        nadawcaId: { type: integer }
        odbiorcaId: { type: integer }
        tytul: { type: string, nullable: true }
        tresc: { type: string }
        dotyczyZwrotuId: { type: integer, nullable: true }
    ReturnSummaryItemDto:
      type: object
      properties:
        id: { type: integer }
        numerZwrotu: { type: string }
        produkt: { type: string }
        ktoPrzyjal: { type: string }
        ktoPodjalDecyzje: { type: string }
        jakaDecyzja: { type: string }
        uwagiMagazynu: { type: string }
        uwagiHandlowca: { type: string }
        status: { type: string }
    ReturnSummaryStatsDto:
      type: object
      properties:
        total: { type: integer }
        doDecyzji: { type: integer }
        zakonczone: { type: integer }
    ReturnSummaryResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ReturnSummaryItemDto' }
        stats:
          $ref: '#/components/schemas/ReturnSummaryStatsDto'
    WarehouseSearchItemDto:
      type: object
      properties:
        nrZgloszenia: { type: string }
        sn: { type: string, nullable: true }
        model: { type: string, nullable: true }
        klient: { type: string, nullable: true }
        produktId: { type: integer }
        kategoria: { type: string, nullable: true }
    WarehouseIntakeRequest:
      type: object
      required: [nrZgloszenia, model, status, lokalizacja]
      properties:
        nrZgloszenia: { type: string }
        model: { type: string }
        numerSeryjny: { type: string, nullable: true }
        status: { type: string }
        lokalizacja: { type: string }
        czyDawca: { type: boolean }
        uwagi: { type: string, nullable: true }
        czesci:
          type: array
          items: { type: string }
    ForwardToComplaintRequest:
      type: object
      required: [returnId, przekazal, daneKlienta, produkt]
      properties:
        returnId: { type: integer }
        powodKlienta: { type: string, nullable: true }
        uwagiMagazynu: { type: string, nullable: true }
        uwagiHandlowca: { type: string, nullable: true }
        przekazal: { type: string }
        daneKlienta:
          $ref: '#/components/schemas/ComplaintCustomerDto'
        produkt:
          $ref: '#/components/schemas/ComplaintProductDto'
    ComplaintCustomerDto:
      type: object
      required: [imie, nazwisko, adres]
      properties:
        imie: { type: string }
        nazwisko: { type: string }
        email: { type: string, nullable: true }
        telefon: { type: string, nullable: true }
        adres:
          $ref: '#/components/schemas/ComplaintAddressDto'
    ComplaintAddressDto:
      type: object
      properties:
        ulica: { type: string, nullable: true }
        kod: { type: string, nullable: true }
        miasto: { type: string, nullable: true }
    ComplaintProductDto:
      type: object
      required: [nazwa]
      properties:
        nazwa: { type: string }
        nrFaktury: { type: string, nullable: true }
        nrSeryjny: { type: string, nullable: true }
    ReturnListResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/ReturnListItemDto' }
            page: { type: integer }
            pageSize: { type: integer }
            totalItems: { type: integer }
            totalPages: { type: integer }
        message: { type: string, nullable: true }
    ReturnDetailsResponse:
      type: object
      properties:
        success: { type: boolean }
        data: { $ref: '#/components/schemas/ReturnDetailsDto' }
        message: { type: string, nullable: true }
    ReturnActionsResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: array
          items: { $ref: '#/components/schemas/ReturnActionDto' }
        message: { type: string, nullable: true }
    MessagesResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: array
          items: { $ref: '#/components/schemas/MessageDto' }
        message: { type: string, nullable: true }
    WarehouseSearchResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: array
          items: { $ref: '#/components/schemas/WarehouseSearchItemDto' }
        message: { type: string, nullable: true }